requirements.txt
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
exceptiongroup==1.0.1
idna==3.4
iniconfig==1.1.1
multidict==6.0.2
packaging==21.3
pluggy==1.0.0
pyparsing==3.0.9
pytest==7.2.0
PyYAML==6.0
requests==2.28.1
six==1.16.0
tomli==2.0.1
urllib3==1.26.12
vcrpy==4.2.1
wrapt==1.14.1
yarl==1.8.1
